Refactor: Organize NavigationBar component folder - Prepare for documentation in future commits.
