+++ client/src/components/NavigationBar/SectionMenus/NotificationCard/NotificationCardTypes.ts
